requirements.txt
Django==4.2.7
django_extensions==4.1
paramiko==3.3.1
RouterOS-api==0.21.0
librouteros==3.4.1
netmiko==4.6.0
python-decouple==3.8
waitress==3.0.2
requests==2.32.5
django-apscheduler==0.7.0
gunicorn==23.0.0
whitenoise==6.9.0
python-dateutil==2.9.0.post0
pyTelegramBotAPI==4.29.1